DAG longest path question
The input consists of a sequence of one or more DAGs. Each DAG D is represented by an adjacency list.
The input is terminated by a line consisting of a single zero.

The output is a sequence of lines, one for each input DAG. Each line contains a single integer indicating the length of the longest path starting from vertex 0.