AVL trees
For a given adjacency matrix of a rooted tree you need to decide whether this tree can be labeled to become an avl tree. 
Vertices of the tree do not have keys but numbered from 0 to n-1 in order to write the adjacency matrix.

The answer for every block should consist of the height of the tree and True/False (True if the tree is an avl tree).